--- src/app/tab2/tab2.page.html
@@ -1,22 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-title>
-      Tab 2
-    </ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">Tab 2</ion-title>
-    </ion-toolbar>
-  </ion-header>
-
-  <app-explore-container name="Tab 2 page"></app-explore-container>
-
-  <!-- Display the data here -->
-  <div *ngFor="let item of data | keyvalue">
-    {{item.key}}: {{item.value}}
-  </div>
-</ion-content>
