+++ src/app/tabs/tabs-routing.module.ts
@@ -12,8 +12,8 @@ const routes: Routes = [
         loadChildren: () => import('../tab1/tab1.module').then(m => m.Tab1PageModule)
       },
       {
-        path: 'tab2',
-        loadChildren: () => import('../tab2/tab2.module').then(m => m.Tab2PageModule)
+        path: 'timeline',
+        loadChildren: () => import('../timeline/timeline.module').then(m => m.TimelinePageModule)
       },
       {
         path: 'tab3',
