+++ src/app/tabs/tabs.page.html
@@ -6,9 +6,9 @@
       <ion-label>Tab 1</ion-label>
     </ion-tab-button>
 
-    <ion-tab-button tab="tab2" href="/tabs/tab2">
-      <ion-icon aria-hidden="true" name="ellipse"></ion-icon>
-      <ion-label>Tab 2</ion-label>
+    <ion-tab-button tab="timeline" href="/timeline">
+      <ion-icon aria-hidden="true" name="hourglass-outline"></ion-icon>
+      <ion-label>Timeline</ion-label>
     </ion-tab-button>
 
     <ion-tab-button tab="tab3" href="/tabs/tab3">
