+++ src/app/timeline/timeline-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { TimelinePage } from './timeline.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: TimelinePage
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class TimelinePageRoutingModule {}
