+++ src/app/timeline/timeline.page.spec.ts
@@ -1,0 +1,17 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { TimelinePage } from './timeline.page';
+
+describe('TimelinePage', () => {
+  let component: TimelinePage;
+  let fixture: ComponentFixture<TimelinePage>;
+
+  beforeEach(async(() => {
+    fixture = TestBed.createComponent(TimelinePage);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
