+++ src/app/timeline/timeline.page.ts
@@ -2,11 +2,12 @@ import { Component, OnInit } from '@angular/core';
 import { FirestoreService } from '../../services/database';
 
 @Component({
-  selector: 'app-tab2',
-  templateUrl: 'tab2.page.html',
-  styleUrls: ['tab2.page.scss'],
+  selector: 'app-timeline',
+  templateUrl: './timeline.page.html',
+  styleUrls: ['./timeline.page.scss'],
 })
-export class Tab2Page implements OnInit {
+export class TimelinePage implements OnInit {
+
   data: any;
 
   constructor(private firestoreService: FirestoreService) {}
